Move controller code into its own package
