+++ cmd/cluster-olm-operator/main.go
@@ -1,38 +1,23 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"errors"
 	goflag "flag"
-	"fmt"
-	"io/fs"
 	"os"
-	"path/filepath"
-	"strings"
 
-	configv1 "github.com/openshift/api/config/v1"
-	operatorv1 "github.com/openshift/api/operator/v1"
 	"github.com/openshift/library-go/pkg/controller/controllercmd"
 	"github.com/openshift/library-go/pkg/controller/factory"
-	"github.com/openshift/library-go/pkg/operator/deploymentcontroller"
-	"github.com/openshift/library-go/pkg/operator/events"
-	"github.com/openshift/library-go/pkg/operator/staticresourcecontroller"
 	"github.com/openshift/library-go/pkg/operator/status"
 	"github.com/openshift/library-go/pkg/operator/v1helpers"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	"golang.org/x/text/cases"
-	"golang.org/x/text/language"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/apimachinery/pkg/util/yaml"
 	"k8s.io/component-base/cli"
 	utilflag "k8s.io/component-base/cli/flag"
-	"k8s.io/klog/v2"
 
 	"github.com/openshift/cluster-olm-operator/pkg/clients"
+	"github.com/openshift/cluster-olm-operator/pkg/controller"
 	"github.com/openshift/cluster-olm-operator/pkg/version"
 )
 
@@ -50,11 +35,11 @@ func newRootCommand() *cobra.Command {
 		Use:   "cluster-olm-operator",
 		Short: "OpenShift Cluster OLM Operator",
 	}
-	cmd.AddCommand(newOperatorCommand())
+	cmd.AddCommand(newStartCommand())
 	return cmd
 }
 
-func newOperatorCommand() *cobra.Command {
+func newStartCommand() *cobra.Command {
 	cmd := controllercmd.NewControllerCommandConfig(
 		"cluster-olm-operator",
 		version.Get(),
@@ -71,10 +56,10 @@ func runOperator(ctx context.Context, cc *controllercmd.ControllerContext) error
 		return err
 	}
 
-	cb := controllerBuilder{
-		assetsFS: os.DirFS("/operand-assets"),
-		cl:       cl,
-		cc:       cc,
+	cb := controller.Builder{
+		Assets:            os.DirFS("/operand-assets"),
+		Clients:           cl,
+		ControllerContext: cc,
 	}
 
 	controllers, relatedObjects, err := cb.BuildControllers("catalogd", "operator-controller", "rukpak")
@@ -97,7 +82,7 @@ func runOperator(ctx context.Context, cc *controllercmd.ControllerContext) error
 		controllerList = append(controllerList, controller)
 	}
 
-	upgradableConditionController := newStaticUpgradeableConditionController(
+	upgradeableConditionController := controller.NewStaticUpgradeableConditionController(
 		"OLMStaticUpgradeableConditionController",
 		cl.OperatorClient,
 		cc.EventRecorder.ForComponent("OLMStaticUpgradeableConditionController"),
@@ -119,7 +104,7 @@ func runOperator(ctx context.Context, cc *controllercmd.ControllerContext) error
 
 	cl.StartInformers(ctx)
 
-	for _, c := range append(controllerList, upgradableConditionController, clusterOperatorController) {
+	for _, c := range append(controllerList, upgradeableConditionController, clusterOperatorController) {
 		go func(c factory.Controller) {
 			defer runtime.HandleCrash()
 			c.Run(ctx, 1)
@@ -128,158 +113,4 @@ func runOperator(ctx context.Context, cc *controllercmd.ControllerContext) error
 
 	<-ctx.Done()
 	return nil
-}
-
-func newStaticUpgradeableConditionController(name string, operatorClient *clients.OperatorClient, eventRecorder events.Recorder, prefixes []string) factory.Controller {
-	c := staticUpgradeableConditionController{
-		name:           name,
-		operatorClient: operatorClient,
-		prefixes:       prefixes,
-	}
-
-	return factory.New().WithSync(c.sync).WithSyncDegradedOnError(operatorClient).WithInformers(operatorClient.Informer()).ToController(name, eventRecorder)
-}
-
-type staticUpgradeableConditionController struct {
-	name           string
-	operatorClient *clients.OperatorClient
-	prefixes       []string
-}
-
-func (c staticUpgradeableConditionController) sync(ctx context.Context, _ factory.SyncContext) error {
-	logger := klog.FromContext(ctx).WithName(c.name)
-	logger.V(4).Info("sync started")
-	defer logger.V(4).Info("sync finished")
-
-	opSpec, _, _, err := c.operatorClient.GetOperatorState()
-	if err != nil {
-		return err
-	}
-	if opSpec.ManagementState != operatorv1.Managed {
-		return nil
-	}
-
-	updateStatusFuncs := make([]v1helpers.UpdateStatusFunc, 0, len(c.prefixes))
-	for _, prefix := range c.prefixes {
-		updateStatusFuncs = append(updateStatusFuncs, v1helpers.UpdateConditionFn(operatorv1.OperatorCondition{
-			Type:   fmt.Sprintf("%sUpgradeable", prefix),
-			Status: operatorv1.ConditionTrue,
-		}))
-	}
-
-	if _, _, updateErr := v1helpers.UpdateStatus(ctx, c.operatorClient, updateStatusFuncs...); updateErr != nil {
-		return updateErr
-	}
-
-	return nil
-}
-
-func replaceImageHook(placeholder string, desiredImageEnvVar string) deploymentcontroller.ManifestHookFunc {
-	return func(spec *operatorv1.OperatorSpec, deployment []byte) ([]byte, error) {
-		replacer := strings.NewReplacer(placeholder, os.Getenv(desiredImageEnvVar))
-		newDeployment := replacer.Replace(string(deployment))
-		return []byte(newDeployment), nil
-	}
-}
-
-type controllerBuilder struct {
-	assetsFS fs.FS
-	cl       *clients.Clients
-	cc       *controllercmd.ControllerContext
-}
-
-func (b *controllerBuilder) BuildControllers(subDirectories ...string) (map[string]factory.Controller, []configv1.ObjectReference, error) {
-	var (
-		controllers    = map[string]factory.Controller{}
-		relatedObjects []configv1.ObjectReference
-		errs           []error
-	)
-
-	titler := cases.Title(language.English)
-
-	for _, subDirectory := range subDirectories {
-		var staticResourceFiles []string
-		namePrefix := strings.ReplaceAll(titler.String(subDirectory), "-", "")
-		if err := fs.WalkDir(b.assetsFS, subDirectory, func(path string, d fs.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if d.IsDir() {
-				return nil
-			}
-			if filepath.Ext(path) != ".yaml" && filepath.Ext(path) != ".yml" {
-				return nil
-			}
-
-			manifestData, err := fs.ReadFile(b.assetsFS, path)
-			if err != nil {
-				errs = append(errs, fmt.Errorf("error reading assets file %q: %w", path, err))
-				return nil
-			}
-
-			var manifest unstructured.Unstructured
-			if err := yaml.NewYAMLOrJSONDecoder(bytes.NewReader(manifestData), 4096).Decode(&manifest); err != nil {
-				errs = append(errs, fmt.Errorf("error parsing manifest for file %q: %w", path, err))
-				return nil
-			}
-
-			manifestGVK := manifest.GroupVersionKind()
-			restMapping, err := b.cl.RESTMapper.RESTMapping(manifestGVK.GroupKind(), manifestGVK.Version)
-			if err != nil {
-				errs = append(errs, fmt.Errorf("error looking up RESTMapping for file %q, gvk %v: %w", path, manifestGVK, err))
-				return nil
-			}
-			relatedObjects = append(relatedObjects, configv1.ObjectReference{
-				Group:     restMapping.GroupVersionKind.Group,
-				Resource:  restMapping.Resource.Resource,
-				Namespace: manifest.GetNamespace(),
-				Name:      manifest.GetName(),
-			})
-
-			if manifestGVK.Kind == "Deployment" && manifestGVK.Group == "apps" {
-				controllerName := fmt.Sprintf("%sDeployment%s",
-					namePrefix,
-					strings.ReplaceAll(titler.String(manifest.GetName()), "-", ""),
-				)
-				controllers[controllerName] = deploymentcontroller.NewDeploymentController(
-					controllerName,
-					manifestData,
-					b.cc.EventRecorder.ForComponent(controllerName),
-					b.cl.OperatorClient,
-					b.cl.KubeClient,
-					b.cl.KubeInformerFactory.Apps().V1().Deployments(),
-					nil,
-					[]deploymentcontroller.ManifestHookFunc{
-						replaceImageHook("${CATALOGD_IMAGE}", "CATALOGD_IMAGE"),
-						replaceImageHook("${OPERATOR_CONTROLLER_IMAGE}", "OPERATOR_CONTROLLER_IMAGE"),
-						replaceImageHook("${RUKPAK_IMAGE}", "RUKPAK_IMAGE"),
-						replaceImageHook("${KUBE_RBAC_PROXY_IMAGE}", "KUBE_RBAC_PROXY_IMAGE"),
-					},
-				)
-				return nil
-			}
-
-			staticResourceFiles = append(staticResourceFiles, path)
-			return nil
-		}); err != nil {
-			return nil, nil, err
-		}
-
-		if len(staticResourceFiles) > 0 {
-			controllerName := fmt.Sprintf("%sStaticResources", namePrefix)
-			controllers[controllerName] = staticresourcecontroller.NewStaticResourceController(
-				controllerName,
-				func(name string) ([]byte, error) { return fs.ReadFile(b.assetsFS, name) },
-				staticResourceFiles,
-				b.cl.ClientHolder(),
-				b.cl.OperatorClient,
-				b.cc.EventRecorder.ForComponent(controllerName),
-			)
-		}
-	}
-	if len(errs) > 0 {
-		return nil, nil, fmt.Errorf("error building controllers: %w", errors.Join(errs...))
-	}
-	return controllers, relatedObjects, nil
 }
