+++ pkg/controller/builder_test.go
@@ -1,0 +1,9 @@
+package controller
+
+import (
+	"testing"
+)
+
+func TestControllerNameForObject(t *testing.T) {
+	// TODO: Add test cases.
+}
